+++ sol/contracts/libraries/SpotFeeLib.sol
@@ -1,9 +1,9 @@
 // SPDX-License-Identifier: AGPL-3.0-only - (c) AirCarbon Pte Ltd - see /LICENSE.md for Terms
 pragma solidity 0.8.5;
 
-import "../libraries/StructLib.sol";
+import { StructLib } from "../libraries/StructLib.sol";
 
-import "../facets/StFeesFacet.sol";
+import { StFeesFacet } from "../facets/StFeesFacet.sol";
 
 library SpotFeeLib {
 	event SetFeeTokFix(uint256 tokTypeId, address indexed ledgerOwner, uint256 fee_tokenQty_Fixed);
@@ -28,10 +28,13 @@ library SpotFeeLib {
 		address[] calldata ledgerOwner,
 		StructLib.SetFeeArgs[] memory a
 	) public {
-		uint len = tokTypeId.length;
-		require(len == ledgerOwner.length && len == a.length, "setFee_TokTypeBatch: argument array lengths does not match");
+		uint256 len = tokTypeId.length;
+		require(
+			len == ledgerOwner.length && len == a.length,
+			"setFee_TokTypeBatch: argument array lengths does not match"
+		);
 
-		for(uint i = 0; i < len; i++) {
+		for (uint256 i = 0; i < len; i++) {
 			setFee_TokType(ld, std, globalFees, tokTypeId[i], ledgerOwner[i], a[i]);
 		}
 	}
@@ -52,7 +55,7 @@ library SpotFeeLib {
 		require(a.ccy_perMillion == 0, "ccy_perMillion unsupported for token-type fee");
 		require(a.ccy_mirrorFee == false, "ccy_mirrorFee unsupported for token-type fee");
 
-		StructLib.FeeStruct storage feeStruct = globalFees; // v2.TODO 
+		StructLib.FeeStruct storage feeStruct = globalFees; // v2.TODO
 		if (ledgerOwner != address(0x0)) {
 			StructLib.initLedgerIfNew(ld, ledgerOwner);
 
@@ -100,9 +103,9 @@ library SpotFeeLib {
 		StructLib.FeeStruct storage globalFees,
 		StructLib.SetFeeCcyTypeBatchArgs[] calldata params
 	) public {
-		uint len = params.length;
+		uint256 len = params.length;
 
-		for(uint i = 0; i < len; i++) {
+		for (uint256 i = 0; i < len; i++) {
 			StructLib.SetFeeCcyTypeBatchArgs calldata param = params[i];
 			setFee_CcyType(ld, ctd, globalFees, param.ccyTypeId, param.ledgerOwner, param.feeArgs);
 		}
